refactor: move logic to helper file
